Tighten up handling of the specific commits to benchmark
This fixes a cluster of related bugs:

- Currently, if two commits are tested in quick succession, each with the same
  base, the base commit will be benchmarked twice. This moves the check to see
  whether the data already exists later, so the second time should not run,
  saving time.

- Currently, when re-running the same commit, the raw results are replaced, but
  the derived results (comparison tables and plots) are not re-generated. This
  detects whether the data for a commit already exists and deletes the entire
  directory, which will force a complete regeneration downstream.

- When the git `fork` or `ref` input is set to something that doesn't exist, the
  git clone of CPython fails, with a completely inscrutable error message that
  looks sort of like a timeout. This detects a broken `cpython` checkout and
  displays a more helpful error message.

This works be adding a `force` flag to the `_benchmarking` and `_pystats` jobs.
By default, the `head` (the commit the user specifies) is always forced. The
`base` is never forced. In other words, `force` is an implementation detail
and not an option presented to the user.

The implementation of this is cumbersome. Github Actions don't have a way for a
step to say "quit this job, but don't fail". So the step has to set a
flag (`should_run`) and then every subsequent step has to check it:

```
if: ${{ steps.should_run.outputs.should_run != 'false' }
```

This does not fix issues related to changing the tip of the branch while the
workflow is running. That kind of thing would be much harder to properly code
against, and I think we currently should deal with that by saying "don't do
that" -- we haven't actually seen that trip anyone up yet.